requirements.txt
clickhouse_connect
pyspark
apache-airflow-providers-apache-spark==2.1.3